+++ spec/support/defra_ruby_validators.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+DefraRubyValidators.configure do |configuration|
+  configuration.companies_house_api_key = ENV["COMPANIES_HOUSE_API_KEY"]
+end
