+++ spec/support/vcr.rb
@@ -1,5 +1,10 @@
 # frozen_string_literal: true
 
+# Stubbing HTTP requests
+require "webmock/rspec"
+# Auto generate fake responses for web-requests
+require "vcr"
+
 VCR.configure do |c|
   c.cassette_library_dir = "spec/cassettes"
   c.hook_into :webmock
